refactor: move fields to widgets layer
