+++ src/widgets/fields/index.ts
@@ -1,0 +1,2 @@
+export { default as GameField } from "./ui/GameField";
+export { default as WordsPicker } from "./ui/WordsPicker";
